mem file: fix eof logic and allow seek beyond end of file
eof logic was changed to match the behaviour of feof. The eof
flag is set when an attempt is made to read bytes beyond the
last byte in the 'file'. The eof flag is reset when writing or
seeking.

It is now possible to seek beyond the end of the 'file'. Also, when
as seek beyond the end of the file is followed by a write then
the file size is increased to equal the file position and then the
new data is written.