avidmxfinfo: support files with missing ec label
Files generated by Avid Airspeed were missing required essence
container labels in both the partition packs as well as the
Preface.
Code no longer assumes that the essence container label is present.
The picture coding label is checked if the essence container label
is not present
The picture coding label is added to the AvidMXFInfo struct