fix RGALayout type
The RGBALayout type is a fixed array with length 8 and does use an
array header. Tha 'var' in the len column in the mxf spec. should be
understood to mean that the number of active components is variable.

Removed RGBALayoutComponent type and added a RGBLayout type.

RGBLayout getters and setters were modifed to set and get a RGBLayout
rather than an RGBLayoutComponent.