add avl tree impl for data model set defs to improve speed
A speed improvement of ~16% was observed when using avidmxfinfo.

The tree is initialised to not accept duplicates and therefore
that check was removed from mxf_check_data_model

The tree was not used for the DataModel::itemDefs because it was
found that it was more expensive than a list for the reading
header metadata case.