file: factor out stream (non-seekable) file handling
Moved stream file handling to a new MXF file structure. Stream files
are wrapped and a stream position is held in memory and updated
when reading or writing.

Changed the check for seek support to check for regular file mode.
A regular file does not include pipes or stdin/stdout/stderr when
not redirected.

Seeking forwards by more than 1 is now supported for streams by calling
mxf_skip, when reading, or mxf_write_zeros when writing.

Added regression test for stdin.