app: add function to check issues and ignore them when reading events
The archive mxf writer had a bug which resulted in Sequence sets in
descriptive metadata event tracks to miss the required
StructuralComponents property.

This update will allow that file issue to be ignored when reading
the event information. The new function can be used to check
whether the file contains the issue.