archive: fix extra vtr event tracks when matching fails
The VTR errors have a timecode read from the RS422 interface. This
timecode needs to be matched with a timecode read through the SDI
interface so that it can be converted into a position.

Not all VTR errors can be matched with a position and therefore there
could be less events stored in the VTR error event track. This could
result in one or more empty event tracks and the issue that the
Sequence set in the event tracks missing the required StructuralCOmponents
property.

This update counts the number of VTR errors that can be matched before
deciding how many event tracks are needed.

The check for at least the first 100 errors matched has been removed.